+++ src/Core/Algorithms/Field/CMakeLists.txt
@@ -32,7 +32,6 @@ SET(Algorithms_Field_SRCS
   InterfaceWithCleaver2Algorithm.cc
   CalculateNodeLocationFrequencyAlgorithm.cc
   RefineTetMeshLocallyAlgorithm.cc
-  CalculateInsideWhichFieldAlgorithm.cc
 )
 
 SET(Algorithms_Field_HEADERS
@@ -42,7 +41,6 @@ SET(Algorithms_Field_HEADERS
   CalculateNodeLocationFrequencyAlgorithm.h
   RefineTetMeshLocallyAlgorithm.h
   share.h
-  CalculateInsideWhichFieldAlgorithm.h
 )
 
 SCIRUN_ADD_LIBRARY(Algorithms_Field
